Implemented connected-component after k-means

+++ findClosestCentroid.cpp
@@ -2,8 +2,8 @@
 
 int findClosestCentroid(Mat* image, pxLocation* centroids, pxLocation* pxLocaiton, int N) {
 
-	int distance;
-	int min = findDistance(image, centroids[0], *pxLocaiton);
+	float distance;
+	float min = findDistance(image, centroids[0], *pxLocaiton);
 	int closestCentroidIndex = 0;
 
 	for (int i = 1; i < N; i++) {
